requirements.txt
Flask~=3.0.3
flask_pydantic~=0.12.0
flask-socketio~=5.3.6

mysqlclient==2.1.0
SQLAlchemy[mysql]~=2.0.31

python-dotenv~=1.0.0
bcrypt~=4.1.3
PyJWT~=2.8.0

pydantic[email]~=2.8.2
pydantic-extra-types~=2.9.0

phonenumbers~=8.13.40

requests~=2.32.3
websockets~=12.0
pandas~=2.2.2